requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-ai-contentsafety==1.0.0
azure-core==1.34.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
dnspython==2.7.0
fastapi==0.115.12
flake8==7.2.0
fonttools==4.58.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
Jinja2==3.1.6
jose==1.0.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mccabe==0.7.0
mongomock==4.3.0
mongomock-motor==0.0.36
motor==3.7.1
numpy==2.2.6
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pillow==11.2.1
pluggy==1.6.0
pycodestyle==2.13.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyflakes==3.3.2
pymongo==4.13.0
pyparsing==3.2.3
pytest==8.3.5
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
seaborn==0.13.2
sentinels==1.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
